requirements.txt
PyRSS2Gen==1.1
python-frontmatter==1.0.0
filehash==0.1.dev5